Payment Types ticket 3

+++ BangazonAPI/Controllers/PaymentTypeController.cs
@@ -52,9 +52,9 @@ namespace BangazonAPI.Controllers
                             customerId = reader.GetInt32(reader.GetOrdinal("CustomerId"))
 
                         };
-                        
-                            paymentTypes.Add(currentpaymentType);
-                        
+
+                        paymentTypes.Add(currentpaymentType);
+
 
                     }
                     reader.Close();
@@ -173,10 +173,24 @@ namespace BangazonAPI.Controllers
                     conn.Open();
                     using (SqlCommand cmd = conn.CreateCommand())
                     {
-                        cmd.CommandText = @"DELETE FROM PaymentType WHERE Id = @id";
-                        
-                    cmd.Parameters.Add(new SqlParameter("@id", id));
+                        //FIND any instances of connections to other tables
+                        cmd.CommandText = @"SELECT[Order].PaymentTypeId AS 'Order PmtType Id' FROM[Order] WHERE[Order].PaymentTypeId = @id";
+                        cmd.Parameters.Add(new SqlParameter("@id", id));
+                        SqlDataReader reader = cmd.ExecuteReader();
 
+
+                        if (reader.Read())
+                        {
+                            //throw error
+                            throw new Exception("Cannot Delete This One");
+                        }
+
+                        else
+                        {
+                            //DELETE IT
+                            cmd.CommandText = @"DELETE FROM PaymentType WHERE Id=@id";
+                        }
+                        reader.Close();
                         int rowsAffected = cmd.ExecuteNonQuery();
                         if (rowsAffected > 0)
                         {
